parisc: Add TOC button support
Add support for an emulated TOC/NMI button.

Almost all PA-RISC machines have either a button that is labeled with
'TOC' or a BMC/GSP function to trigger a TOC.  TOC is a non-maskable
interrupt that is sent to the processor.  This can be used for
diagnostic purposes like obtaining a stack trace/register dump or to
enter KDB/KGDB in Linux.

This patch adds support for such an emulated TOC button.

This patch writes the CPU registers into PIM (processor internal
memmory) for later analysis.  It uses the CPU "shadow registers", the
IAOQ-back and IASQ-back values are provided by qemu in r24/r25.  This
patch uses the new qemu aritificial opcode "getshadowregs" (0xfffdead2)
which restores the original values of the shadow registers.

To trigger a TOC, switch to the "qemu monitor" with Ctrl-A C, and type
in the command "nmi".  After the TOC started the OS-debugger, exit the
qemu monitor with Ctrl-A C.

Signed-off-by: Helge Deller <deller@gmx.de>