parisc: Allow 16 CPUs
Move the HPA for MEMORY_HPA out of the address space of the 16th CPU,
and fix the PDC_COPROC_CFG PDC function to report functional FPUs for
all configured CPUs. The documentation specifies that cr10 can only
report for 8 CPUs, but that seems outdated as both HP-UX 10 and HP-UX 11
accepts this change.

Signed-off-by: Helge Deller <deller@gmx.de>