parisc: Add opt/console fw_cfg option to select default console
By default the firmware will choose the (virtual) graphics card
as firmware console if the graphics card is available.
If not it defaults to serial console.

In some configurations it's useful to select the serial console
als default output device even if the graphics card is available.
For that, this patch adds the opt/console fw_cfg qemu option.

Qemu option usage to switch the firmare to use the serial or graphics
default output:

-fw_cfg opt/console,string=serial
-fw_cfg opt/console,string=graphics

Signed-off-by: Helge Deller <deller@gmx.de>