parisc/lasips2: Fix lasips2_kbd_in() to return early without input
lasips2_kbd_in() is used to read input from an emulated PS/2 keyboard.
Since there is no interrupt support durig PDC/IODC we need to poll the
PS/2 status flags to see if a new key was pressed.  In order to be able
to poll regularily, exit the loop to read already pressed chars as soon
as possible. This can be done by calling handle_16() with register ah =
0x11 to check if keys are queued up and exit if none are available.

This patch fixes the problem that keys couldn't be read when using a
graphical output (artist graphics). This patch fixes palo and other
bootloaders to be able to use input from the emulated PS/2 keyboard.

Signed-off-by: Helge Deller <deller@gmx.de>