parisc: Switch serial ports 1 and 2
On real hardware, Linux detects the UART which is built-into LASI as serial port #1,
and the UART which is built-into DINO as serial port #2.

But to now qemu and SeaBIOS used the UART on DINO as serial port #1,
which breaks qemu's possibility to use the "-serial" option to
pass-through host serial ports to the guest.

Fix this by swapping the addresses of PORT_SERIAL1 and PORT_SERIAL2.

Signed-off-by: Helge Deller <deller@gmx.de>