parisc: SEABIOS_HPPA_VERSION 7
Fixes boot of the Debian-12 installation CD-ROM

The size of the ramdisc of Debian-12 is 17123328 bytes and occupies 8361
CD-ROM sectors (of 2048 bytes each).  As such the palo bootloader tries to
read 8361 sectors, which fails with DISK_RET_EBADTRACK error because the
emulated LSI SCSI driver only supports (due to hardware limitation)
reading 8192 sectors at once.
Fix it in the SeaBIOS firmware by limiting the sectors which can be read
at once to 8192-16 sectors.  The palo bootloader will then simply do
additional calls to read the remaining sectors.
This firmware bug didn't show up with Debian-11 because that ramdisc
was smaller and below the 8192 sector limit.

Signed-off-by: Helge Deller <deller@gmx.de>