parisc: Prepare for addition of a HP 715 machine
On the HP 715 machine the base address of the LASI controller is
at 0xf0100000, so introduce a constant for this address.

The firmware end address needs to be reduced to not overlap with
the LASI addresses.

Introduce LASI_XXX constants which reflects the offsets of the
various devices of the LASI base HPA.

Adjust the serial port addressess accordingly. We can not use hard
coded HPAs, because on 715 the LASI serial ports are at different
addresses than on B160L.

Signed-off-by: Helge Deller <deller@gmx.de>